Fix name formatting for Asia/Dushanbe in 1991 March-September
We were only applying "slash" formatting (where the format specifies
daylight and standard time separately) for zone lines which referred to a named rule.
Asia/Dushanbe has a zone line like this:

5:00 1:00 +05/+06 1991 Sep  9  2:00s

Now admittedly specifying a format of +05/+06 for a line which is fixed as daylight is a bit odd, but it should still work.
